fix(ci): drop vitest from barcode-1d runtime
