+++ code/packages/typescript/barcode-1d/BUILD_windows
@@ -10,4 +10,4 @@ cd ..\upc-a && npm ci --quiet
 npm ci --quiet
 cargo build --release --manifest-path Cargo.toml
 if exist target\release\barcode_1d_native_node.dll copy target\release\barcode_1d_native_node.dll barcode_1d_native_node.node
-npx vitest run --coverage
+node --test tests/barcode-1d.test.js
