+++ code/packages/typescript/barcode-1d/package.json
@@ -11,8 +11,8 @@
     "barcode_1d_native_node.node"
   ],
   "scripts": {
-    "test": "vitest run",
-    "test:coverage": "vitest run --coverage"
+    "test": "node --test tests/barcode-1d.test.js",
+    "test:coverage": "node --test tests/barcode-1d.test.js"
   },
   "dependencies": {
     "@coding-adventures/barcode-layout-1d": "file:../barcode-layout-1d",
@@ -23,10 +23,6 @@
     "@coding-adventures/itf": "file:../itf",
     "@coding-adventures/paint-instructions": "file:../paint-instructions",
     "@coding-adventures/upc-a": "file:../upc-a"
-  },
-  "devDependencies": {
-    "vitest": "^3.1.1",
-    "@vitest/coverage-v8": "^3.1.1"
   },
   "license": "MIT"
 }
