--- code/packages/typescript/barcode-1d/vitest.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    coverage: {
-      provider: "v8",
-      include: ["index.js"],
-    },
-  },
-});
